thêm tiêu chí, tính năng chấm điểm

+++ database/migrations/2024_10_24_061809_create_danh_sach_tieu_chis_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->text('ten_tieu_chi');
             $table->float('diem_toi_da');
             $table->float('diem_toi_thieu')->default(0);
-            $table->tinyInteger('he_so');
+            $table->float('he_so');
             $table->boolean('trang_thai')->default(false);
 
             // $table->timestamps();
